[2단계 - 상품 관리 기능] 하디(전동혁) 미션 제출합니다 (#297)
* refactor: 헤더 값을 파싱하여, Id를 알아내도록 변경

* feat: JdbcProductDao 구현

* refactor: Entity명 변경

* config: 데이터베이스 설정 정보 추가

* feat: 데이터베이스 테이블 설계

* refactor: 관리자 페이지 수정

* refactor: 홈페이지 수정

* feat: 회원가입 페이지 구현

* style: final keyword 작성

* style: final keyword 작성

* feat: 로그인 페이지 구현

* feat: 커스텀 예외 구현

* feat: 사용자 엔티티 및 디티오 구현

* feat: 사용자 DAO 구현

* feat: 사용자 서비스 구현

* feat: 사용자 뷰 컨트롤러 구현

* feat: 사용자 api 컨트롤러 구현

* feat: 간단한 시저 암호화, 복호화 기능구현

* feat: 커스텀 예외 핸들링 기능 추가

* feat: 어드민 권한 기능 추가

* refactor: 모달 스타일 변경

* style: 주석제거

* refactor: 테이블 수정

* feat: cart 페이지 작성

* refactor: 홈페이지 금액 '원' 텍스트 추가

* style: 주석 제거

* refactor: 어드민 페이지 '원' 텍스트 추가

* style: 개행 제거

* refactor: productDao 구현체 수정

* refactor: final keyword 추가

* style: import 패키지 수정

* feat: 비밀번호 암호화, 복호화 카이사르 알고리즘 기능 구현

* feat: 관리자 권한 예외 구현

* refactor: 관리자 페이지 패키지 경로 수정

* refactor: 홈페이지 패키지 경로 수정

* refactor: 상수 변경

* feat: base64 리졸버 구현

* feat: base64 인터셉터 구현

* feat: base64 ArgumentResolver, Interceptor 추가

* feat: Auth 커스텀 어노테이션 구현

* refactor: 패키지 경로 변경

* feat: CartDao 구현

* feat: CartService 구현

* feat: CartPage 접속 기능 구현

* feat: CartController 기능 구현

* feat: Cart 엔티티 구현

* feat:AdminAccess 예외처리 추가

* feat: 패스워드 불일치 예외 구현

* feat: 태스트 yml 및 테스트 테이블 설계

* test: productDaoTest 작성

* test: memberDaoTest 작성

* test: cartDaoTest 작성

* test: ProductManagementService test 작성

* test: MemberService test 작성

* test: CartService test 작성

* refactor: 패키지 경로 수정

* refactor: base64 인증에 따른 헤더 변경

* test: MemberPageControllerTest 작성

* test: MemberControllerTest 작성

* refactor: 패키지 경로 수정

* refactor: base64 인증에 따른 헤더 변

* test: CartPageControllerTest 기능 구현

* test: CartControllerTest 기능 구현

* feat: service 트랜잭션 적용

* docs: Readme 업데이트

* refactor: super method 제거

* refactor: 패키지 경로 customExceptions -> customexceptions로 자바 컨벤션 지키도록 변경

* refactor: delete -> deleteById 메서드 명 변경

* refactor: orElseThrow를 사용하여 변경

* style: 주석 제거

* refactor: Mock 어노테이션을 이용

* style: 개행 제거

* style: 안쓰는 임포트 제거

* refactor: cartDao 기능 수정

* refactor: cartId를 사용하도록 리팩터링

* feat: cartItemDto 설계

* refactor: api 스펙 변경

* refactor: api 스펙 변경

* refactor: 검증 변경

* refactor: api 변경에 따른 테스트 수정

* feat: UnauthorizedException 예외 추가

* feat: AuthService 기능 구현

* refactor: 인터셉터, ArgumentResolver 빈으로 등록

* docs: README 업데이트

* refactor: js 수정

* feat: 패스워드 확인 로직 추가

* refactor: 패스워드 확인 로직 변경

* refactor: 인터셉터 빈 등록에 의한 테스트코드 변경

* refactor: 이메일 중복 확인 DAO에서 검증

* test: 커스텀 예외로 인한 테스트 변경

* feat: domain validation 추가

* feat: domain test 추가

* style: 주석 제거

* feat: logging 추가

+++ src/main/resources/static/js/admin.js
@@ -46,9 +46,20 @@ form.addEventListener('submit', (event) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const createProduct = (product) => {
+
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
         method: 'post',
         url: '/admin/products',
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
         data: {
             name: product.name,
             imageUrl: product.imageUrl,
@@ -57,7 +68,9 @@ const createProduct = (product) => {
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
+        alert(error.response.data.message);
         console.error(error);
+        window.location.reload();
     });
 };
 
@@ -65,9 +78,19 @@ const createProduct = (product) => {
 const updateProduct = (product) => {
     const { id } = product;
 
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
-        method: "patch",
+        method: "put",
         url: '/admin/products/' + id,
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
         data: {
             name: product.name,
             imageUrl: product.imageUrl,
@@ -76,18 +99,33 @@ const updateProduct = (product) => {
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
+        alert(error.response.data.message);
         console.error(error);
+        window.location.reload();
     });
 };
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const deleteProduct = (id) => {
+
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
         method: "delete",
-        url: '/admin/products/' + id
+        url: '/admin/products/' + id,
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        }
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
+        alert(error.response.data.message);
         console.error(error);
+        window.location.reload();
     });
 };
